+++ docs/CHANGES.md
@@ -14,6 +14,9 @@ All changes are logged here. These are auto-generated using the Merge & Complete
 
 
 
+
+## 2025-11-23 - Gavin Wallis - Platform Feature Fix
+- Ensure monitoring integration commits endpoint docs during creation.
 ## 2025-11-23 - Gavin Wallis - New Service
 - Add monitoring endpoint: nginx (network - 10.66.33.112)
 ## 2025-11-23 - Gavin Wallis - Platform Feature Fix